Create README.md(3)
I have a burning desire to find a mentor in IT/ CYBER SECURITY.  i'm a HUNGRY beginner and im looking to change the course of my life onto something with purpose and distance myself from ppl who aren't elevating me. I'm a good student and even places to start would be appreciated!  Thanks!